jni: Fix JNI floating point tests
This patch fixes floating point test cases JNITest by changing "jfloat" and
"jdouble" types to float and double, respectively. This ensures that generated
native code uses floating point registers to pass the return values.

Reported-by: Joonas Reynders <joonas.reynders@iki.fi>

+++ include/vm/jni.h
@@ -36,8 +36,8 @@ typedef uint16_t jchar;
 typedef int16_t jshort;
 typedef int32_t jint;
 typedef int64_t jlong;
-typedef uint32_t jfloat;
-typedef uint64_t jdouble;
+typedef float jfloat;
+typedef double jdouble;
 
 typedef jint jsize;
